Adjacency lists

These functions are all-vertices version of neighbors() and incident().

as_adjlist(graph, mode = 3L)

as_inclist(graph, mode = 3L)

Arguments

graph
IGraph object.

mode
Edge type to count; 1: OUT, 2: IN, 3: ALL.

Details

as_adjlist(): Convert a graph to a list of adjacent vertices.

as_inclist(): Convert a graph to a list of incident edges.

See also