Centrality measures

The betweenness centrality of an edge is the number of geodesics going through it.

Usage

edge_betweenness(graph, directed = is_directed(graph), weights = numeric(0))

edge_betweenness_subset(
  graph,
  directed = is_directed(graph),
  eids = integer(0),
  weights = numeric(0),
  from = integer(0),
  to = integer(0)
)

Source

https://igraph.org/c/doc/igraph-Structural.html#centrality-measures

Arguments

graph

An igraph_ptr object.

directed

A logical value, whether to consider directed paths. Ignored for undirected graphs.

weights

A numeric vector of edge weights; TRUE to use Eattr(graph, "weight").

eids

An integer vector of edge IDs.

from

An integer vector of vertex IDs.

to

An integer vector of vertex IDs.

Value

edge_betweenness() returns a numeric vector of edge betweenness for each edge in the graph.

edge_betweenness_subset() is a variant of edge_betweenness() that computes the betweenness using a subset of paths between from and to.

Examples

g = graph_tree(5L)
edge_betweenness(g)
#> [1] 3 1 2 2
edge_betweenness_subset(g, from = Vsource(g), to = Vsink(g))
#> [1] 2 1 1 1