Deterministic Graph Generators

Deterministic Graph Generators

Usage

graph_create(edges, n = 0L, directed = TRUE)

graph_star(n, mode = 0L, center = 1L)

graph_lattice(
  dim,
  nei = 1L,
  directed = FALSE,
  mutual = FALSE,
  circular = FALSE
)

graph_ring(n, directed = FALSE, mutual = FALSE, circular = TRUE)

graph_tree(n, children = 2L, mode = 0L)

graph_full(n, directed = FALSE, mutual = FALSE)

graph_famous(name)

Source

https://igraph.org/c/doc/igraph-Generators.html

Arguments

edges

The edges to add, the first two elements are the first edge, etc.

n

Integer, the number of vertices in the graph.

directed

Boolean

mode

Integer;

igraph_star_mode_t {0: OUT, 1: IN, 2: MUTUAL, 3: UNDIRECTED}

igraph_tree_mode_t {0: OUT, 1: IN, 2: UNDIRECTED}

center

Id of the vertex which will be the center of the graph.

dim

Vector giving the sizes of the lattice in each of its dimensions.

nei

Integer distance within which two vertices will be connected.

mutual

Boolean

circular

Boolean

children

Integer, the number of children of a vertex in the tree.

name

{Bull, Chvatal, Coxeter, Cubical, Diamond, Dodecahedral, Dodecahedron, Folkman, Franklin, Frucht, Grotzsch, Heawood, Herschel, House, HouseX, Icosahedral, Icosahedron, Krackhardt_Kite, Levi, McGee, Meredith, Noperfectmatching, Nonline, Octahedral, Octahedron, Petersen, Robertson, Smallestcyclicgroup, Tetrahedral, Tetrahedron, Thomassen, Tutte, Uniquely3colorable, Walther, Zachary}

See also